imx-gpu-viv: Update install for improved packaging design
The old packaging model stripped the version number from the
filename, added a backend qualifier, and removed any symbolic
links. This had to be matched in the install with logic to
undo the changes.

The design is now greatly simplified by preserving all library
filenames as-is, with backend-specific libraries placed in
sub-folders.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>