imx-gpu-viv: skip QA check for libgl-imx package
libgl-imx package provides libGL, which has runtime dependencies on X11
libraries. In case of Wayland-only builds, this triggers the QA error on
missing RDEPENDS, since they are not provided by any package.

Skip this QA check for libgl-imx package, as it is not distributed for
Wayland-only builds, and has dependencies properly resolved when X11
builds are used.

Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>